Moving getIoFromCli() function into io.ts file.
