+++ src/bin/main.ts
@@ -1,7 +1,7 @@
 import { cliParams } from "../lib/cli/arg-parser.js";
-import { getIoFromCli } from "../lib/cli/index.js";
 import config from "../lib/config/index.js";
 import Engine from "../lib/engine/engine.js";
+import { getIoFromCli } from "../lib/io/io.js";
 import { Database } from "../lib/model/database.js";
 import Slack from "../lib/services/slack.js";
 import { AttachableError, SimpleError } from '../lib/util/errors.js';
