+++ src/bin/run-once.ts
@@ -1,6 +1,6 @@
 import { cliParams } from "../lib/cli/arg-parser.js";
-import { getIoFromCli } from "../lib/cli/index.js";
 import Engine from "../lib/engine/engine.js";
+import { getIoFromCli } from "../lib/io/io.js";
 import { Database } from "../lib/model/database.js";
 
 const io = getIoFromCli();
